# sycl: add usage of enqueue_functions extension