… ggml/src/ggml-sycl/cpy.cpp
… ggml/src/ggml-sycl/element_wise.cpp
… ggml/src/ggml-sycl/getrows.cpp
… ggml/src/ggml-sycl/ggml-sycl.cpp
… ggml/src/ggml-sycl/gla.cpp
… ggml/src/ggml-sycl/im2col.cpp
… ggml/src/ggml-sycl/mmq.cpp
… ggml/src/ggml-sycl/mmvq.cpp
… ggml/src/ggml-sycl/norm.cpp
… ggml/src/ggml-sycl/rope.cpp
… ggml/src/ggml-sycl/softmax.cpp
… ggml/src/ggml-sycl/tsembd.cpp
… ggml/src/ggml-sycl/wkv.cpp