@CISC 2025-06-17
> This is going to conflict with the PRs #14158 and #14181. @CISC do you have a rough estimation of when your PRs could be merged? We can wait to merge this one or help with the rebase if needed.

They are ready once @ggerganov updates Metal implementation.