@s-Nick 2025-06-20
@CISC  Thank you for the update.
@qnixsynapse  I think it should be easy enough for you to rebase on this PR, but if you need help, ping me. I am happy to help rebase the SYCL backend.