@NeoZhangJianyu 2025-09-08
@s-Nick 
There are users find there are obviously memory leak with this feature.

In a inference, the memory increase 80M in one interact session (22s).
In workload test, 16GB GPU will be fulled after about 2 hours.

It happens on 2025.0- 2025.2.

I suggest to revert this PR since it includes big issue.
After this feature is fixe in new SYCL version, this PR could be restored.

Here is the log in my test.

```
15:18:09.972,    0,  N/A, 6008.30
15:18:10.972,    0,  N/A, 6012.74
15:18:11.972,    0,  N/A, 6017.19
15:18:12.972,    0,  N/A, 6021.64
15:18:13.972,    0,  N/A, 6026.08
15:18:14.972,    0,  N/A, 6030.53
15:18:15.972,    0,  N/A, 6034.97
15:18:16.972,    0,  N/A, 6039.36
15:18:17.972,    0,  N/A, 6043.74
15:18:18.972,    0,  N/A, 6047.94
15:18:19.972,    0,  N/A, 6052.20
15:18:20.972,    0,  N/A, 6056.45
15:18:21.972,    0,  N/A, 6060.65
15:18:22.972,    0,  N/A, 6064.84
15:18:23.972,    0,  N/A, 6065.28
15:18:24.972,    0,  N/A, 6065.28
15:18:25.972,    0,  N/A, 6065.28
15:18:26.972,    0,  N/A, 6066.41
15:18:27.972,    0,  N/A, 6070.67
15:18:28.972,    0,  N/A, 6074.80
15:18:29.972,    0,  N/A, 6079.06
15:18:30.972,    0,  N/A, 6083.44
15:18:31.972,    0,  N/A, 6087.39

```